% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{ang.deg}
\alias{ang.deg}
\title{Angle of a plane with \emph{z}}
\usage{
ang.deg(XYZplane)
}
\arguments{
\item{XYZplane}{sample from a point cloud. Matrix n x 3, with columns containing x, y and z coordinates, respectively.}
}
\value{
angle between the plane's normal vector and the \emph{z} axis \code{c(0,0,1)} (in degrees).
}
\description{
Calculates the angle between a plane's normal vector and the \emph{z} axis \code{c(0,0,1)}. The plane here is a point cloud sample.
}

