% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{angle}
\alias{angle}
\title{Angle between two vectors}
\usage{
angle(a, b)
}
\arguments{
\item{a, b}{numerical vectors of same length}
}
\value{
angle between \emph{a} and \emph{b} (in degrees).
}
\description{
Calculates the angle between two vectors.
}
