% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{availablePointMetrics}
\alias{availablePointMetrics}
\title{Print available point metrics}
\usage{
availablePointMetrics()
}
\value{
vector with available metrics names.
}
\description{
print available point metrics - for usage in \code{\link{pointMetrics}}.
}
\examples{
m = availablePointMetrics()
print(m)
}
