% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{balls}
\alias{balls}
\title{Rough noise removal based on 3D spheres}
\usage{
balls(xyz.tree, ball.rad = 0.025, np.min = 2, sec.filt = T,
  lball.rad = 0.05, min.ncov = 3)
}
\arguments{
\item{xyz.tree}{tree point cloud, \emph{xyz} matrix}

\item{ball.rad}{spheres radii for first cleaning}

\item{np.min}{minimum number of points per sphere - spheres with less points will be considered noise and thus removed}

\item{sec.filt}{apply noise filter a second time?}

\item{lball.rad}{spheres radii for second cleaning (should be larger than ball.rad) - only used if \code{sec.filt == TRUE}}

\item{min.ncov}{minimum number of points per sphere for the second cleaning - only used if \code{sec.filt == TRUE}}
}
\value{
tree point cloud without rough noise
}
\description{
removes isolated points from a point cloud
}

