% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{center.zero}
\alias{center.zero}
\title{Centralize of \emph{xy} to zero}
\usage{
center.zero(xyz.cloud)
}
\arguments{
\item{xyz.cloud}{point cloud - \emph{xyz} matrix}
}
\value{
point cloud with new coordinates
}
\description{
reassigns the x and y coordinates of a point cloud considering zero as its center
}

