% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{circlefit}
\alias{circlefit}
\title{Least squares circle fit}
\usage{
circlefit(xp, yp, fast = FALSE, c0 = T)
}
\arguments{
\item{xp}{x coordinates}

\item{yp}{y coordinates}

\item{fast}{if \code{TRUE} skips the optimization step}

\item{c0}{if \code{TRUE}, centers x and y to zero}
}
\value{
vector object containing x, y center coordinates, circle radius and sum of squared errors
}
\description{
Fits a circle to a set of points - adapted from the \code{pracma} package.
}

