% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{clip.XY}
\alias{clip.XY}
\title{Circular point cloud clip}
\usage{
clip.XY(cloud, rad = 1.5, center = c(0, 0))
}
\arguments{
\item{cloud}{point cloud - \emph{xyz} matrix}

\item{rad}{circle radius}

\item{center}{x and y center coordinates for the circle}
}
\value{
point cloud of all points inside the specified circle
}
\description{
clips a region of a point cloud in circular shape
}
