% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{cloud.col}
\alias{cloud.col}
\title{Point cloud coloring by height}
\usage{
cloud.col(xyz.cloud, pal = rainbow, n = 10, quantile = F)
}
\arguments{
\item{xyz.cloud}{point cloud - \emph{xyz} matrix}

\item{pal}{color pallete - defaults to \code{rainbow}}

\item{n}{number of height intervals to color differently}

\item{quantile}{divide point cloud by height quantiles? - If FALSE, divides the point cloud in equal height intervals}
}
\value{
plots \emph{xyz.cloud} using \code{rgl} and with the specified color pallete
}
\description{
colors a point cloud by height intervals
}

