% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{cube.space}
\alias{cube.space}
\title{3D sample points cloud construction}
\usage{
cube.space(xyz.tree, l = 0.03, systematic = T, n = 3, iterate = 1)
}
\arguments{
\item{xyz.tree}{tree point cloud, \emph{xyz} matrix}

\item{l}{minimum distance between sample points}

\item{systematic}{distribute the points systematically as a voxel grid? TRUE or FALSE}

\item{n, iterate}{only used if \code{systematic == FALSE}. The higher those values are, the "denser" is the resuling sample points point cloud}
}
\value{
3D point cloud of sample points - \emph{xyz} matrix
}
\description{
creates sample points in the 3D space randomly or systematically distributed
}
