% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{cyl.dists}
\alias{cyl.dists}
\title{Distance from the cylinder surface}
\usage{
cyl.dists(ang.rad, P)
}
\arguments{
\item{ang.rad}{vector containing the 5 cylinder parameters: \emph{rho, theta, phi, alpha and radius} - as in \code{\link{cyl.parameters}} output}

\item{P}{cylinder point cloud - \emph{xyz} matrix}
}
\value{
vector of distances from every point to the cylinder's surface
}
\description{
Calculates the distances from every point in a point cloud to its model cylinder
}

