% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{cyl.fit}
\alias{cyl.fit}
\title{Sum of squared distances from a cylinder surface}
\usage{
cyl.fit(ang.rad, P)
}
\arguments{
\item{ang.rad}{vector containing the 5 cylinder parameters: \emph{rho, theta, phi, alpha and radius} - as in \code{\link{cyl.parameters}} output}

\item{P}{cylinder point cloud - \emph{xyz} matrix}
}
\value{
sum of squared distances from \emph{P} to the cylinder surface with \emph{ang.rad} parameters
}
\description{
Calculates the sum of squared distances from a cylinder surface
}

