% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{cyl.iso}
\alias{cyl.iso}
\title{Isolation of cylinder from stem point cloud}
\usage{
cyl.iso(stem, H = 1.3, bound = 0.5, method = "circle")
}
\arguments{
\item{stem}{stem point cloud - \emph{xyz} matrix}

\item{H}{cylinder's/circle's mean height}

\item{bound}{height interval from H, i.e. H +/- bound}

\item{method}{\emph{"circle"} or \emph{"cylinder"}}
}
\value{
vector of circle or cylinder parameters and its residual sum of squares related to the extracted \emph{stem}'s section
}
\description{
Takes out a stem section in a specified interval and gets its circle or cylinder parameters
}

