% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{cyl.parameters}
\alias{cyl.parameters}
\title{Estimate cylinder parameters from a point cloud}
\usage{
cyl.parameters(P, init = NULL, opt.method = "Nelder-Mead")
}
\arguments{
\item{P}{cylinder point cloud - \emph{xyz} matrix}

\item{init}{(optional) initial guess of cylinder parameters}

\item{opt.method}{optimization method - passed on to \code{\link[stats]{optim}}}
}
\value{
\code{\link[stats]{optim}} output containing the optimal cylinder parameters of \emph{P}, respectively: \emph{rho, theta, phi, alpha and radius}
}
\description{
Estimates the cylinder parameters of a point cloud
}

