% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{cyl.vectors}
\alias{cyl.vectors}
\title{Cylinder direction vectors extraction}
\usage{
cyl.vectors(ang.rad)
}
\arguments{
\item{ang.rad}{vector containing the 5 cylinder parameters: \emph{rho, theta, phi, alpha and radius} - as in \code{\link{cyl.parameters}}}
}
\value{
list containing the vectors, respectively: \emph{n, a, Q, n^theta and n^phi bar}
}
\description{
Extracts the vectors that define a cylinder's direction based on its five main parameters - as in \code{\link{cyl.parameters}} output
}
