% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cylinderFit}
\alias{cylinderFit}
\title{Point cloud cylinder fit}
\usage{
cylinderFit(las, method = "ransac", n = 5, inliers = 0.9, p = 0.95,
  max_angle = 30)
}
\arguments{
\item{las}{\code{LAS} object.}

\item{method}{method for estimating the cylinder parameters. Currently available: \code{"nm"}, \code{"irls"}, \code{"ransac"} and \code{"bf"}.}

\item{n}{\code{integer} - number of points selected on every RANSAC iteration.}

\item{inliers}{\code{numeric} - expected proportion of inliers among stem segments' point cloud chunks.}

\item{max_angle}{\code{numeric} - applicable when \code{method == "bf"}. It's the maximum angle a point cloud's axis can deviate from an absolute vertical axis (Z = c(0,0,1) ), in degrees.}

\item{conf}{\code{numeric} - confidence level.}
}
\description{
Fits a cylinder on a set of 3D points.
}
