% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fastPointMetrics.available}
\alias{fastPointMetrics.available}
\title{Print available point metrics}
\usage{
fastPointMetrics.available(enable = ENABLED_POINT_METRICS$names)
}
\arguments{
\item{enable}{optional \code{integer} or \code{character} vector. Indices or names of the metrics you want to enable globally. Enabled metrics are calculated everytime you run \code{\link{fastPointMetrics}}.}
}
\value{
\code{character} vector of available metrics.
}
\description{
print available point metrics - used in \code{\link{fastPointMetrics}}.
}
\examples{
m = fastPointMetrics.available()
print(m)
}
