% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sets.R
\name{fit_IRTLS}
\alias{fit_IRTLS}
\title{Iterated reweighted total least squares cylinder fit for stem isolation}
\usage{
fit_IRTLS(trunk, c.len = 0.5, max.rad = 0.5, s.height = 1, speed.up = T,
  opt.method = "Nelder-Mead")
}
\arguments{
\item{trunk}{output point cloud from a pre-filtering method. Matrix with 3 columns: x, y and z coordinates, respectively}

\item{c.len}{length of stem segments to fit (in meters)}

\item{max.rad}{maximum radius accepted as estimate (in meters)}

\item{s.height}{starting height for acquiring baseline cylinder parameters (in meters)}

\item{speed.up}{TRUE or FALSE, if TRUE takes no more than 500 points to fit a cylinder, randomly selected (applicable for large point clouds)}

\item{opt.method}{optimization method, passed to optim (R base function)}
}
\value{
list of length 2: /n $stem.points = single stem point cloud, matrix with 3 columns, x, y and z, respectively /n $cylinders = matrix displaying the optimal parameters of all cylinders fitted along the stem
}
\description{
Iterated reweighted total least squares cylinder fit for stem isolation
}

