% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sets.R
\name{fit_RANSAC_circle}
\alias{fit_RANSAC_circle}
\title{RANSAC circle fit stem isolation}
\usage{
fit_RANSAC_circle(trunk, l.int = 0.5, cut.rad = 0.01, n = 15,
  p = 0.8, P = 0.99)
}
\arguments{
\item{trunk}{output point cloud from a pre-filtering method. Matrix with 3 columns: x, y and z coordinates, respectively}

\item{l.int}{length of stem segments to fit (in meters)}

\item{cut.rad}{value to add to the radius, removing all points outside a range of radius+cut.rad from the point cloud (in meters)}

\item{n}{number of points to take in every RANSAC iteration}

\item{p}{estimated proportion of inliers (stem points)}

\item{P}{confidence level}
}
\value{
list of length 2: /n $stem.points = single stem point cloud, matrix with 3 columns, x, y and z, respectively /n $circles = matrix displaying the optimal parameters of all circles fitted along the stem
}
\description{
RANSAC circle fit stem isolation
}
