% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sets.R
\name{fit_RANSAC_cylinder}
\alias{fit_RANSAC_cylinder}
\title{RANSAC cylinder fit stem isolation}
\usage{
fit_RANSAC_cylinder(trunk, c.len = 0.5, h.init = 1, max.rad = 0.5,
  timesN = 2, opt.method = "Nelder-Mead")
}
\arguments{
\item{trunk}{output point cloud from a pre-filtering method. Matrix with 3 columns: x, y and z coordinates, respectively}

\item{c.len}{length of stem segments to fit (in meters)}

\item{h.init}{height to start the fitting procedure}

\item{max.rad}{maximum radius accepted as estimate (in meters)}

\item{timesN}{factor that multiplies N, for the number of iterations of the RANSAC, N=36 as default}

\item{opt.method}{optimization method, passed to optim (R base function)}
}
\value{
list of length 2: /n $stem.points = single stem point cloud, matrix with 3 columns, x, y and z, respectively /n $cylinders = matrix displaying the optimal parameters of all cylinders fitted along the stem
}
\description{
RANSAC cylinder fit stem isolation
}

