% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{FL}
\alias{FL}
\title{Surface flatness}
\usage{
FL(XYZplane)
}
\arguments{
\item{XYZplane}{point cloud - \emph{xyz} matrix}
}
\value{
surface flatness - ranging from 0 (non-flat) to 1 (perfectly flat)
}
\description{
calculates the flatness of a point cloud
}

