% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{hough}
\alias{hough}
\title{Circle hough transformation for a single circle}
\usage{
hough(raster, rad = c(0.025, 0.5), pixel_size = 0.025, min.val = 0.1,
  Plot = F, img.prefix = "", ...)
}
\arguments{
\item{raster}{output from \code{\link{makeRaster}}}

\item{rad}{lower and upper limits of radii to survey}

\item{pixel_size}{pixel side length, in meters}

\item{min.val}{minimum pixel density or frequency that applies for testing}

\item{Plot}{if TRUE, saves a .png file showing all sample points generated for the input raster}

\item{img.prefix}{file name for \emph{Plot}}

\item{...}{further arguments passed on to \code{\link[graphics]{plot}}}
}
\value{
object of class 'list' with 3 compartments: \describe{\item{$centers}{
matrix with 4 columns, each row contains the x and y center coordinates,circle radius and number of 'votes' (overlapping peripheral circles)}
\item{$images}{list of arrays, each one containing the votes per raster cell for one iterated radius}
\item{$circles}{list of xy coordinates of all peripheral circles tested, each compartment contains a matrix with xy point coordinates from circles with same radius (the ones used in 'images')
}}
}
\description{
estimates the circle parameters for a point cloud using the Hough transformation
}
\seealso{
\code{\link{makeRaster}}
}
