% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maping_funcs.R
\name{hough_plot}
\alias{hough_plot}
\title{Hough transformation plot-wise}
\usage{
hough_plot(raster, rad = c(0.025, 0.5), nRad = 50, nAng = 90,
  min.val = 0.1, votes = 3, Plot = F, img.prefix = "", ...)
}
\arguments{
\item{raster}{output from \code{\link{makeRaster}}}

\item{rad}{interval of radii to fit circles}

\item{nRad}{number of radii to test in between \code{rad}}

\item{nAng}{number of points in a circle}

\item{min.val}{minimum pixel density value to be included in the circle fitting procedure}

\item{votes}{minimum number of votes for a circle to be included in the output}

\item{Plot}{if TRUE, saves a .png file showing all sample points generated for the input raster}

\item{img.prefix}{file name for Plot}

\item{...}{arguments passed to \code{\link{plot}}}
}
\value{
matrix of all circle centers found
}
\description{
Applies the hough transformation on TLS plot, returning ALL circles found
}

