% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{HT_base_filter}
\alias{HT_base_filter}
\title{Tree base Hough transformation filter}
\usage{
HT_base_filter(XYZmat, z.int = NULL, rad.inf = 2, cell.size = 0.025,
  min.val = 0.3, Plot = F)
}
\arguments{
\item{XYZmat}{single tree point cloud - \emph{xyz} matrix}

\item{z.int}{optional - height interval to take the reference cylinder. If not specified, the height interval adopted is from 5\% to 10\% of the tree's total height}

\item{rad.inf}{inflation factor to multuply the radius. All points (in the entire point cloud) outside a range of \emph{rad.inf * radius} from the reference cylinder's center will be deleted}

\item{cell.size}{pixel size for the Hough transformation}

\item{min.val}{passed on to \code{\link{hough}}}

\item{Plot}{plot the reference tree segment? TRUE or FALSE}
}
\value{
vector of length 5, containing the upper and lower height limits of the reference cylinder, xy coordinates of circle center and its radius
}
\description{
identification of reference cylinder at the tree's base for filtering outliers using the Hough transformation
}
\seealso{
\code{\link{hough}}
}
