% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{IRLS}
\alias{IRLS}
\title{Iterated Reweighted Least Squares cylinder fit}
\usage{
IRLS(P, init = NULL, max.iter = 10, Plot = F, speed.up = F,
  opt.method = "Nelder-Mead")
}
\arguments{
\item{P}{cylinder point cloud - \emph{xyz} matrix}

\item{init}{(optional) initial guess of cylinder parameters}

\item{max.iter}{maximum number of iterations}

\item{Plot}{if TRUE plots the x and y cylinder coordinates highlighting the 25\% lowest weight points}

\item{speed.up}{for large point clouds (> 500 points) - if TRUE, performs optimization over 500 randomly selected points only}

\item{opt.method}{optimization method - passed on to \code{\link[stats]{optim}}}
}
\value{
list containing the cylinder optimal parameters, its sum of squared distances from the point cloud, number of iterations and point weights, respectively
}
\description{
Estimates the optimal cylinder parameters of a point cloud using Iterated Reweighted Least Squares
}

