% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{las2xyz}
\alias{las2xyz}
\title{Extracts XYZ data from \code{LAS}}
\usage{
las2xyz(las)
}
\arguments{
\item{las}{\code{LAS} object}
}
\value{
XYZ \code{matrix} object
}
\description{
Extracts the XYZ slots from a \code{LAS} object into a \code{matrix}
}
