% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{makeRaster}
\alias{makeRaster}
\title{Create raster from a point cloud}
\usage{
makeRaster(XYZmat, cell.size = 0.01, density = T, image = T)
}
\arguments{
\item{XYZmat}{point cloud - \emph{xyz} matrix}

\item{cell.size}{pixel size}

\item{density}{if \code{TRUE}, each pixel will display point density values, if \code{FALSE}, frequency values will be used}

\item{image}{if \code{TRUE}, plots the raster as an image}
}
\value{
object of clas 'list' with 4 compartments: \describe{\item{$x}{cell breaks in x}
\item{$y}{cell breaks in y}
\item{$z}{matrix representing each individual raster cell with their respective density or frequency values}
\item{$classes}{x and y intervals explicitly identified according to the values in the z compartment}}
}
\description{
extracts a raster (\emph{xy}) containing density or frequency information from a xyz a point cloud
}

