% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{map.eigen.knn}
\alias{map.eigen.knn}
\title{Tree mapping algorithm: KNN point geometry}
\usage{
map.eigen.knn(
  max_curvature = 0.1,
  max_verticality = 10,
  max_mean_dist = 0.1,
  max_d = 0.5,
  min_h = 1.5,
  max_h = 3,
  min_n = 100
)
}
\arguments{
\item{max_curvature}{\code{numeric} - maximum curvature (from 0 to 1) accepted when filtering a point neighborhood.}

\item{max_verticality}{\code{numeric} - maximum deviation of a point neighborhood's orientation from an absolute vertical axis ( Z = c(0,0,1) ), in \emph{degrees} (from 0 to 90).}

\item{max_mean_dist}{\code{numeric} - maximum average distance between points tolerated in a neighborhood.}

\item{max_d}{\code{numeric} - maximum diameter tolerated for a tree segment.}

\item{min_h}{\code{numeric} - minimum height of a point cluster to consider it as a tree.}

\item{min_n}{\code{numeric} - lowest point count accepted in a tree's point cluster.}
}
\description{
This function is meant to be used inside \code{\link{treeMap}}. It applies a KNN filter to select points with specific neighborhood features. For more details on geometry features, check out \code{\link{fastPointMetrics}}.
}
