% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{map.pick}
\alias{map.pick}
\title{Tree mapping algorithm: pick trees manually}
\usage{
map.pick(min_h = NULL, max_h = NULL)
}
\arguments{
\item{min_h, max_h}{minimum/maximum height thresholds to filter the point cloud before plotting it in \code{rgl}.}
}
\description{
This function is meant to be used inside \code{\link{treeMap}}. It opens a new \code{rgl} plot on which the user can point the tree locations by clicking.
}
