% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maping_funcs.R
\name{map_HT}
\alias{map_HT}
\title{Map stem coordinates in a TLS point cloud using the Hough transformation}
\usage{
map_HT(xyz.cloud, z.lim = 1:2, min.votes = 2, pixel.size = 0.05,
  min.den = 0.1, map.rad = 1.5, Plot = T, ...)
}
\arguments{
\item{xyz.cloud}{forest plot point cloud - \emph{xyz} matrix}

\item{z.lim}{height interval from which a slice of the point cloud will be taken}

\item{min.votes}{minimum number of votes for a circle to be used in the method}

\item{pixel.size}{pixel side length - passed to \code{\link{makeRaster}}}

\item{min.den}{minimum pixel density value to be surveyed in \code{\link{hough_plot}}}

\item{map.rad}{radius to aggregate circle centers to calculate a stem's coordinate}

\item{Plot}{if TRUE plots the cloud's slice with the estimated stem cooridinates}

\item{...}{arguments passed to \code{\link{plot}}}
}
\value{
xy matrix with all estimated stem coordinates
}
\description{
Outputs the approximate stem positions in a TLS point cloud, estimated by circular patterns found through the hough transformation
}

