% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maping_funcs.R
\name{map_SD}
\alias{map_SD}
\title{Map stem coordinates in a TLS point cloud through spectral decomposition filter}
\usage{
map_SD(xyz.cloud, z.lim = 1:2, ppm = 2000, k = 30, flat.min = 0.9,
  ang.tol = 20, map.rad = 1.5, Plot = T, ...)
}
\arguments{
\item{xyz.cloud}{forest plot point cloud - \emph{xyz} matrix}

\item{z.lim}{height interval from which a slice of the point cloud will be taken}

\item{ppm}{points per meter to retain across the largest dimension (x or y)}

\item{k}{number of points to analyze in a neighborhood}

\item{flat.min}{minimum tolerated flatness value to use a point neighborehood in the analysis}

\item{ang.tol}{maximum angle deviation with \emph{z} tolerated to use a point neighborehood in the analysis}

\item{map.rad}{radius to aggregate point neighborhoods to calculate a stem's coordinate}

\item{Plot}{if TRUE plots the cloud's slice with the estimated stem cooridinates}

\item{...}{arguments passed to \code{\link{plot}}}
}
\value{
xy matrix with all estimated stem coordinates
}
\description{
Outputs the approximate stem positions in a TLS point cloud, estimated from surface flatness and verticality obtained through spectral decomposition
}

