% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{neighbours}
\alias{neighbours}
\title{Cover sets neghborhoods assignment}
\usage{
neighbours(xyz.tree, samples3d, d = 0.03, neighborhood = 2)
}
\arguments{
\item{xyz.tree}{tree point cloud, \emph{xyz} matrix}

\item{samples3d}{output point cloud from \code{\link{cube.space}} built for \code{xyz.tree}}

\item{d}{radius of spherical neighborhoods}

\item{neighborhood}{order of neighborhoods to include - 1 means only the direct neighbours, 2 includes the neighbours of first neighbours and so on}
}
\value{
list object in which every compartment contains a cover set neighborhood
}
\description{
Defines cover set spherical neighborhoods for a 3D point cloud based on euclidian distances
}

