% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{neighbours2}
\alias{neighbours2}
\title{Cover sets neghborhoods assignment}
\usage{
neighbours2(xyz.tree, samples3d, neighborhood = 3)
}
\arguments{
\item{xyz.tree}{tree point cloud, \emph{xyz} matrix}

\item{samples3d}{output point cloud from \code{\link{cube.space}} built for \emph{xyz.tree} - the points here must be systematically distributed}

\item{neighborhood}{order of voxel neighborhoods to include - 1 means only the direct neighbour voxels, 2 includes the neighbours of direct neighbours and so on}
}
\value{
list object in which every compartment contains a cover set neighborhood
}
\description{
Defines cover set cubic neighborhoods based on a voxel space
}
