% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nnFilter}
\alias{nnFilter}
\title{Filter isolated points}
\usage{
nnFilter(las, d = 0.05, n = 2)
}
\arguments{
\item{las}{\code{LAS} object.}

\item{d}{\code{numeric} - distance to search for neighboring points.}

\item{n}{\code{integer} - number of points below which a neighborhood (within distance \emph{d}) will be assigned as noise}

\item{max_points}{\code{numeric} - total number of points tolerated per point cloud chunk to apply knn search. If the input \code{LAS} is larger than \code{max_points}, knn search will be performed over point cloud slices (along the Z axis) with \emph{N} up to \code{max_points}.}
}
\value{
\code{LAS} object.
}
\description{
Noise filtering algorithm based on point neighborhood distances.
}
\examples{
file = system.file("extdata", "spruce.laz", package="TreeLS")
tls = readTLS(file)
}
