% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sets.R
\name{pref_HT}
\alias{pref_HT}
\title{Circle hough transformation pre-filter}
\usage{
pref_HT(tree, l.int = 0.5, cell.size = 0.025, min.den = 0.3)
}
\arguments{
\item{tree}{single tree point cloud. Matrix with 3 columns: x, y and z coordinates, respectively}

\item{l.int}{length of segments to split the tree into, from bottom to top (in meters)}

\item{cell.size}{pixel size for creating raster layers (in meters)}

\item{min.den}{minimum point density of pixels to test circles using the hough transformation}
}
\value{
rough trunk point cloud - matrix with 3 columns, x, y and z, respectively.
}
\description{
Circle hough transformation pre-filter (based on Olofsson et al. 2014)
}

