% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sets.R
\name{pref_SD}
\alias{pref_SD}
\title{Spectral decomposition pre-filter}
\usage{
pref_SD(tree, k = 30, flat.min = 0.9, ang.tol = 10, l.int = 0.5,
  freq.ratio = 0.25)
}
\arguments{
\item{tree}{single tree point cloud. Matrix with 3 columns: x, y and z coordinates, respectively}

\item{k}{number of points belonging to a surface to be tested for flatness}

\item{flat.min}{minimum tolerated flatness (between 0 and 1)}

\item{ang.tol}{maximum deviation from a perpendicular angle with the z axis (90 +/- ang.tol, in degrees)}

\item{l.int}{length of segments to split the tree into, from bottom to top (in meters) (for this method it only relates to code speed and to recuce RAM usage)}

\item{freq.ratio}{minimum proportion of number of points in a group, in relation to the largest found group, in order to keep it in the output point cloud (between 0 and 1)}
}
\value{
rough trunk point cloud - matrix with 3 columns, x, y and z, respectively.
}
\description{
Spectral decomposition pre-filter (based on Liang et. al 2012)
}
