% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sets.R
\name{pref_VN}
\alias{pref_VN}
\title{Voxel neighborhoods pre-filter}
\usage{
pref_VN(tree, noise1.rad = 0.05, noise2.rad = 0.1, flat.min = 0.9,
  ang.tol = 10, neighborhood = 4, largest.cov = NULL, axis.dist = 0.5)
}
\arguments{
\item{tree}{single tree point cloud. Matrix with 3 columns: x, y and z coordinates, respectively}

\item{noise1.rad}{sphere radii to use for the first rough noise filtering step (in meters)}

\item{noise2.rad}{sphere radii to use for the second rough noise filtering step (in meters)}

\item{flat.min}{minimum tolerated flatness (between 0 and 1)}

\item{ang.tol}{maximum deviation from a perpendicular angle with the z axis (90 +/- ang.tol, in degrees)}

\item{neighborhood}{order of voxel neighborhood to merge (for voxels of 5 cm)}

\item{largest.cov}{minimum accepeted proportion of points in a cover set, in relation to the largest cover set, to keep it in the output cloud. If =NULL an automated maximum curve detection method is applied.}

\item{axis.dist}{maximum distance from an estimated z axis tolerated to keep a cover set in the output cloud}
}
\value{
rough trunk point cloud - matrix with 3 columns, x, y and z, respectively.
}
\description{
Voxel neighborhoods pre-filter (based on Raumonen et al. 2013)
}

