% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_metrics_methods.R
\name{ptm.knn}
\alias{ptm.knn}
\title{Point metrics algorithm: Nearest Neighborhood metrics}
\usage{
ptm.knn(k = 20, r = 0)
}
\arguments{
\item{k}{\code{numeric} - number of closest points to search per neighborhood.}

\item{r}{\code{numeric} - limited search radius. If \code{r == 0}, no distance limit is applied.}
}
\description{
This function is meant to be used inside \code{\link{fastPointMetrics}}. It calculates metrics from a point's nearest neighborhood (KNN).
}
