% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_metrics_methods.R
\name{ptm.voxel}
\alias{ptm.voxel}
\title{Point metrics algorithm: Voxel metrics}
\usage{
ptm.voxel(d = 0.1, exact = FALSE)
}
\arguments{
\item{d}{\code{numeric} - voxel spacing, in point cloud units.}

\item{exact}{\code{logical} - use exact voxel search? If \code{FALSE}, applies approximate voxel search using integer index hashing, much faster on large point clouds (several million points).}
}
\description{
This function is meant to be used inside \code{\link{fastPointMetrics}}. It calculates metrics per voxel.
}
