% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{RANSAC.circle}
\alias{RANSAC.circle}
\title{RAndom SAmple Consensus circle fit}
\usage{
RANSAC.circle(stem.sec, n = 15, p = 0.8, P = 0.99)
}
\arguments{
\item{stem.sec}{stem section, \emph{xyz} matrix}

\item{n}{number of points to sample on every RANSAC iteration}

\item{p}{estimated proportion of inliers in the dataset}

\item{P}{level of confidence desired}
}
\value{
vector object containing x, y center coordinates, circle radius and sum of squared errors
}
\description{
returns the best circle parameters using the RANSAC algorithm
}
