% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{res.stats}
\alias{res.stats}
\title{Precision and accuracy of estimated diameters}
\usage{
res.stats(rad, d)
}
\arguments{
\item{rad}{vector of estimated radii along a tree's stem}

\item{d}{correspondent diameters to \emph{rad}, measured from the field}
}
\value{
vector containing RMSE, bias and correlation, respectively
}
\description{
extracts RMSE, bias and Pearson's correlation from estimated diameters in relation to their respective field measured values
}

