% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{resetLAS}
\alias{resetLAS}
\title{Reset the properties of a \code{LAS} object}
\usage{
resetLAS(las)
}
\arguments{
\item{las}{\code{LAS} object}
}
\value{
newly assigned \code{LAS} object
}
\description{
Rebuilds the \code{LAS} object based on its data table only - all header params are recalculated
}
