% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{rglAXES}
\alias{rglAXES}
\title{Plot 3D axes from the origin}
\usage{
rglAXES(xyz = c(1, 1, 1), cols = c("red", "green", "blue"), ...)
}
\arguments{
\item{xyz}{length of each axis, x, y and z, respectively}

\item{cols}{color of x, y and z, respectively}

\item{...}{further arguments passed to the \code{rgl.lines} function}
}
\value{
plots 3D axes over the current rgl environment
}
\description{
plots the 3 main axes (\emph{xyz}) starting at \code{c(0,0,0)}
}

