% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{SD.prefilt}
\alias{SD.prefilt}
\title{Spectral Decomposition point cloud filter}
\usage{
SD.prefilt(sec, k, flat.min, ang.tol)
}
\arguments{
\item{sec}{tree section, \emph{xyz} matrix}

\item{k}{number of closest points in a neighborhood over which spectral decomposition is performed}

\item{flat.min}{minimum flatness accepted to keep points in the dataset}

\item{ang.tol}{angle tolerance (in degrees) between \code{sec}'s normal vector and \emph{z}}
}
\value{
filtered tree section
}
\description{
Removes points from the provided point cloud whose neighborhoods don't follow the specified criteria
}
