% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setHeaderTLS}
\alias{setHeaderTLS}
\title{Set LAS Header to TLS applications}
\usage{
setHeaderTLS(las, xfac = 1e-04, yfac = 1e-04, zfac = 1e-04)
}
\arguments{
\item{las}{\code{LAS} object}

\item{xfac, yfac, zfac}{scale factor of XYZ axes}
}
\value{
copy of \code{las} with the altered header
}
\description{
Alters the header of an input \code{LAS} object by increasing the precision of XYZ scale factors when suitable
}
