% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setLAS}
\alias{setLAS}
\title{Resets or creates a \code{LAS} object depending on the input's type}
\usage{
setLAS(cloud, ...)
}
\arguments{
\item{cloud}{object to be converted or reset}

\item{...}{further arguments passed to \code{\link{toLAS}} for table-like inputs}
}
\value{
\code{LAS} object
}
\description{
Resets the input whenever it is a \code{LAS} object, or generates a new \code{LAS} from a table-like input
}
