% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_segmentation_methods.R
\name{sgmt.ransac.circle}
\alias{sgmt.ransac.circle}
\title{Stem segmentation algorithm: RANSAC circle fit}
\usage{
sgmt.ransac.circle(tol = 0.025, n = 10, conf = 0.99, inliers = 0.8)
}
\arguments{
\item{tol}{\code{numeric} - tolerance offset between absolute radii estimates and hough transform estimates.}

\item{n}{\code{integer} - number of points selected on every RANSAC iteration.}

\item{conf}{\code{numeric} - confidence level.}

\item{inliers}{\code{numeric} - expected proportion of inliers among stem segments' point cloud chunks.}
}
\description{
This function is meant to be used inside \code{\link{stemSegmentation}}. It applies a least squares circle fit algorithm in a RANSAC fashion over stem segments. Mode details are given in the sections below.
}
\section{Output Fields}{


\itemize{
\item \code{TreeID}:  unique tree IDs - available only for multiple stems
\item \code{Segment}: stem segment number (from bottom to top)
\item \code{X}, \code{Y}: circle center coordinates
\item \code{Radius}: estimated circles radii
\item \code{Error}: least squares circle fit error
\item \code{AvgHeight}: average height of stem segments
\item \code{N}: number of points in the stem segments
}
}

\references{
Conto, T. ; Olofsson, K. ; Gorgens, E. B. ; Rodriguez, L. C. E. ; Almeida, G. Performance of stem denoising and stem modelling algorithms on single tree point clouds from terrestrial laser scanning. Computers and Electronics in Agriculture, v. 143, p. 165-176, 2017.
}
