% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_segmentation_methods.R
\name{sgt.irls.circle}
\alias{sgt.irls.circle}
\title{Stem segmentation algorithm: Iterated Reweighted Least Squares circle fit}
\usage{
sgt.irls.circle(tol = 0.1, n = 500)
}
\arguments{
\item{tol}{\code{numeric} - tolerance offset between absolute radii estimates and hough transform estimates.}

\item{n}{\code{numeric} - maximum number of points to sample when fitting a stem segment.}
}
\description{
This function is meant to be used inside \code{\link{stemSegmentation}}. It applies a reweighted least squares circle fit algorithm using M-estimators in order to remove outlier effects.
}
