% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_segmentation_methods.R
\name{sgt.irls.cylinder}
\alias{sgt.irls.cylinder}
\title{Stem segmentation algorithm: Iterated Reweighted Least Squares cylinder fit}
\usage{
sgt.irls.cylinder(tol = 0.05, n = 100)
}
\arguments{
\item{tol}{\code{numeric} - tolerance offset between absolute radii estimates and hough transform estimates.}

\item{n}{\code{numeric} - maximum number of points to sample when fitting a stem segment.}
}
\description{
This function is meant to be used inside \code{\link{stemSegmentation}}. It applies a reweighted least squares cylinder fit algorithm using M-estimators and Nelder-Mead optimization in order to remove outlier effects.
}
