% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{shapeFit}
\alias{shapeFit}
\title{Point cloud cylinder/circle fit}
\usage{
shapeFit(
  stem_segment = NULL,
  shape = "circle",
  algorithm = "ransac",
  n = 10,
  conf = 0.95,
  inliers = 0.9,
  n_best = 10,
  z_dev = 30
)
}
\arguments{
\item{stem_segment}{\code{NULL} or a \code{\link[lidR:LAS]{LAS}} object with a single stem segment. When \code{NULL} returns a parameterized function to be used as input in other functions (e.g. \code{\link{tlsInventory}}).}

\item{shape}{\code{character} - either \code{circle} or \code{cylinder}.}

\item{algorithm}{optimization method for estimating the shape parameters. Currently available: \code{"ransac"}, \code{"irls"}, \code{"nm"}, \code{"qr"} (circle only) ,\code{"bf"} (cylinder only).}

\item{n}{\code{integer} - number of points selected on every RANSAC iteration.}

\item{conf}{\code{numeric} - confidence level.}

\item{inliers}{\code{numeric} - expected proportion of inliers among stem segments' point cloud chunks.}

\item{n_best}{\code{integer} - estimate optimal RANSAC parameters as the median of the \code{n_best} estimations with lowest error.}

\item{z_dev}{\code{numeric} - maximum angle deviation for brute force cylinder estimation (\code{bf}), i.e. angle, in degrees (0-90), that a cylinder can be tilted in relation to a perfect vertival axis (\code{Z = c(0,0,1)}).}
}
\description{
Fits a 3D cylinder or 2D circle on a set of 3D points, retrieving the optimized parameters.
}
