% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_methods.R
\name{smp.voxelize}
\alias{smp.voxelize}
\title{Point sampling algorithm: systematic voxel grid}
\usage{
smp.voxelize(spacing = 0.05)
}
\arguments{
\item{spacing}{\code{numeric} - voxel side length, in point cloud units.}
}
\description{
This function is meant to be used inside \code{\link{tlsSample}}. It selects one random point per voxel at a given spatial resolution.
}
