% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{stem.model}
\alias{stem.model}
\title{Plotting a stem model}
\usage{
stem.model(stem.out, cyl.len = NA, col = rainbow, bg = "black",
  alpha = 0.5)
}
\arguments{
\item{stem.out}{output from a stem fitting function - \code{\link{fit_RANSAC_circle}}, \code{\link{fit_RANSAC_cylinder}} or \code{\link{fit_IRTLS}}}

\item{cyl.len}{optional - cylinder length for all stem segments}

\item{col}{color pallete function or color string name to use to color the cylinders}

\item{bg}{background color of the rgl environemnt}

\item{alpha}{alpha value passed on to the \code{\link{ashape3d}} function}
}
\value{
3D stem model of stacked cylinders/circles
}
\description{
plots a stem model of stacked cylinders or circles, depending on the \emph{fitting} routine used to calculate the stem segments
}
\examples{
trunk <- pref_HT(spruce)
stem <- fit_RANSAC_circle(trunk)
obj3d <- stem.model(stem)
rgl.points(spruce, size=1)
}

