% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stemPoints_plot}
\alias{stemPoints_plot}
\title{Plot-wise stem point classification}
\usage{
stemPoints_plot(las, map, hstep = 0.5, max_radius = 0.25,
  hbase = c(1, 2.5), pixel_size = 0.025, min_density = 0.1,
  min_votes = 3)
}
\arguments{
\item{las}{\code{LAS} object}

\item{map}{map of tree positions - output from \code{\link{treeMap}} or \code{\link{treePositions}}}

\item{hstep}{...}

\item{max_radius}{...}

\item{hbase}{...}

\item{pixel_size}{...}

\item{min_density}{...}

\item{min_votes}{...}
}
\value{
\code{LAS} object
}
\description{
Classify stem points through the Hough Transform algorithm - it searches for one stem per tree in a point cloud
}
