% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tlsInventory}
\alias{tlsInventory}
\title{Extract forest inventory metrics from a point cloud}
\usage{
tlsInventory(
  las,
  dh = 1.3,
  dw = 0.5,
  hp = 1,
  d_method = shapeFit(shape = "circle", algorithm = "ransac", n = 15, n_best = 20)
)
}
\arguments{
\item{las}{normalized \code{\link[lidR:LAS]{LAS}} object - single tree or forest plot with stem points marked by \code{\link{stemPoints}}.}

\item{dh}{\code{numeric} - height layer (above ground) from which to estimate stem diameters.}

\item{dw}{\code{numeric} - height layer width.}

\item{hp}{\code{numeric} - height percentile to extract per tree (0-1). Use 1 for top height.}

\item{d_method}{parameterized \code{\link(shapeFit)} function - method to use for diameter estimation.}
}
\description{
Estimation of diameter and height tree-wise.
}
