% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tlsNormalize}
\alias{tlsNormalize}
\title{Normalize a TLS point cloud}
\usage{
tlsNormalize(las, res = 0.5, keepGround = T)
}
\arguments{
\item{las}{\code{LAS} object}

\item{res}{resolution of the DTM used for normalization}

\item{keepGround}{default = \code{TRUE} - if \code{FALSE}, returns a point cloud with ground points removed}
}
\value{
\code{LAS} object
}
\description{
Normalizes a TLS point based on a Digital Terrain Model of the ground points. If the input's ground points are not classified, the \code{\link{csf}} algorithm is applied internally.
}
