% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{toLAS}
\alias{toLAS}
\title{Convert a table-like object to \code{LAS}}
\usage{
toLAS(dataMatrix, namesVector = NULL)
}
\arguments{
\item{dataMatrix}{object to be converted to \code{LAS}}

\item{namesVector}{\code{colnames} of the variables in the \code{data} compartment of the \code{LAS} otput.
If \code{NULL} it ignores any names, considering the first 3 columns as XYZ data and nothing else.}
}
\value{
\code{LAS} object
}
\description{
Generates a \code{LAS} object from a \code{matrix}, \code{data.frame} or similar structure
}
