% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{treeMap.merge}
\alias{treeMap.merge}
\title{Merge nearby trees in a \emph{tree_map}}
\usage{
treeMap.merge(las, d = 0.2)
}
\arguments{
\item{las}{\code{LAS} or \code{data.table} object generated by \code{\link{treeMap}}.}

\item{d}{\code{numeric} - distance criteria to merge}
}
\description{
Check all tree neighborhoods and merge TreeIDs which are too close in a \code{tree_map} object.
}
