% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{treeMap}
\alias{treeMap}
\title{Map tree occurrences from TLS data}
\usage{
treeMap(las, hmin = 1, hmax = 3, hstep = 0.5, pixel = 0.025,
  rad_max = 0.25, min_den = 0.1, min_votes = 3)
}
\arguments{
\item{las}{\code{LAS} object}

\item{hmin}{\code{numeric} - ...}

\item{hmax}{\code{numeric} - ...}

\item{hstep}{\code{numeric} - ...}

\item{pixel}{\code{numeric} - ...}

\item{rad_max}{\code{numeric} - ...}

\item{min_den}{\code{numeric} between 0 and 1 - ...}

\item{min_votes}{\code{integer} - ...}
}
\value{
\code{LAS} object
}
\description{
Estimates tree probability regions from a point cloud based on a Hough Transform circle search
}
