% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{treePoints}
\alias{treePoints}
\title{Classify areas of influence for individual trees}
\usage{
treePoints(las, map, method = trp.voronoi())
}
\arguments{
\item{las}{\code{\link[lidR:LAS]{LAS}} object.}

\item{map}{output from \code{\link{treeMap}} or \code{\link{treeMap.positions}}.}

\item{method}{segmentation algorithm - currently available: \code{\link{trp.voronoi}} and \code{\link{trp.crop}}.}
}
\value{
\code{\link[lidR:LAS]{LAS}} object.
}
\description{
Identify the \code{TreeID} of every point on a \code{LAS} object based on coordinates from a \code{\link{treeMap}}. Tree region segmentation methods are prefixed by \code{trp}.
}
\examples{
file = system.file("extdata", "pine_plot.laz", package="TreeLS")
tls = readTLS(file)
}
