% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{treePositions}
\alias{treePositions}
\title{Get tree XY positions from a TLS-tree map}
\usage{
treePositions(las)
}
\arguments{
\item{las}{\code{LAS} object - \code{treeMap}'s output}
}
\value{
\code{data.frame} of tree IDs and XY coordinates
}
\description{
Estimates tree probability regions from a point cloud based on a Hough Transform circle search
}
