% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_points_methods.R
\name{trp.autoseg}
\alias{trp.autoseg}
\title{Tree points algorithm: auto segmentation.}
\usage{
trp.autoseg()
}
\description{
This function is meant to be used inside \code{\link{treePoints}}. Assign points to a nearest \code{TreeID} in \code{\link{treeMap}} coordinates.
}
