% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_points_methods.R
\name{trp.crop}
\alias{trp.crop}
\title{Tree points algorithm: define tree regions on fixed areas.}
\usage{
trp.crop(l = 1, circle = TRUE)
}
\arguments{
\item{l}{\code{numeric} - circle radius or square side length.}

\item{circle}{\code{logical} - assign \emph{TreeID}s inside circles (TRUE) or squares (FALSE)?}
}
\description{
This function is meant to be used inside \code{\link{treePoints}}. Assign points to a \emph{TreeID} inside circles/squares of fixed area around \code{\link{treeMap}} coordinates.
}
