% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{tukey.estimator}
\alias{tukey.estimator}
\title{Tukey's biweight function}
\usage{
tukey.estimator(errors, b = 5)
}
\arguments{
\item{errors}{residuals from a fitted model}

\item{b}{efficiency constant (5 == 95\%)}
}
\value{
list of length 2: \describe{\item{$Y}{residuals / MAD} \item{$weights}{weights}}
}
\description{
calculates weights for a dataset based on Tukey's biweight function
}
