% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{Vfilter}
\alias{Vfilter}
\title{Height-based point cloud filter}
\usage{
Vfilter(XYZtree, l.int = 0.3, thr = 10000)
}
\arguments{
\item{XYZtree}{point cloud (not necessarily for a whole a tree) - \emph{xyz} matrix}

\item{l.int}{length of height intervals to split the point cloud into}

\item{thr}{maximum density threshold - i.e. maximum amount of points tolerated per height interval}
}
\value{
point cloud with reduced density
}
\description{
reduces a point cloud's density processing different height intervals individually
}
\seealso{
\code{\link{Vsections}}
}
