% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{Vsections}
\alias{Vsections}
\title{Split point cloud into height intervals}
\usage{
Vsections(XYZtree, n.int = 100, l.int = NULL, overlap = NULL, Plot = T,
  units = "m", ...)
}
\arguments{
\item{XYZtree}{point cloud (not necessarily a tree) - \emph{xyz} matrix}

\item{n.int}{number of height intervals (divided by point quantiles)}

\item{l.int}{optional - length of height intervals. If not NULL, splits a point cloud by fixed height intervals, instead of using \emph{n.int}}

\item{overlap}{optional - by how much should the height segments overlap?}

\item{Plot}{create a plot for every point cloud segment? TRUE or FALSE}

\item{units}{if \code{Plot == TRUE}, provide the units of measurement for labelling the plots}

\item{...}{further arguments passed to \code{plot}}
}
\description{
divides a point cloud in many smaller ones, according to height intervals
}

