% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{Xprod}
\alias{Xprod}
\title{Cross product between two vectors}
\usage{
Xprod(a, b)
}
\arguments{
\item{a, b}{numerical vectors of same length}
}
\value{
vector cross product
}
\description{
calculates the cross product between two vectors - not to be confused with \code{crossprod}, which gives the dot product
}
