% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{xyz.rotation.matrix}
\alias{xyz.rotation.matrix}
\title{3D rotation matrix}
\usage{
xyz.rotation.matrix(ax, az, ax2)
}
\arguments{
\item{ax}{angle of 1st rotation -  around the \emph{x} axis (in degrees)}

\item{az}{angle of 2nd rotation -  around the \emph{z} axis (in degrees)}

\item{ax2}{angle of 3rd rotation - around the new \emph{x} axis (in degrees)}
}
\value{
3 x 3 \emph{xyz} rotation matrix
}
\description{
calculates the 3D rotation matrix for a \emph{xyz} dataset
}
