% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pointMetrics.available}
\alias{pointMetrics.available}
\title{Print available point metrics}
\usage{
pointMetrics.available()
}
\value{
vector with available metrics names.
}
\description{
print available point metrics - for usage in \code{\link{pointMetrics}}.
}
\examples{
m = pointMetrics.available()
print(m)
}
