% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{circleFit}
\alias{circleFit}
\title{Point cloud circle fit}
\usage{
circleFit(las, method = "irls", n = 5, inliers = 0.8, p = 0.99,
  n_best = 0)
}
\arguments{
\item{las}{\code{LAS} object.}

\item{method}{method for estimating the circle parameters. Currently available: \code{"qr"}, \code{"nm"}, \code{"irls"} and \code{"ransac"}.}

\item{n}{\code{integer} - number of points selected on every RANSAC iteration.}

\item{inliers}{\code{numeric} - expected proportion of inliers among stem segments' point cloud chunks.}

\item{n_best}{\code{numeric} - Applicable when \code{method == "ransac"}. For \code{n_best == 0}, takes the best RANSAC iteration as estimates, otherwise, it estimates the parameters as the average of \code{n_best} RANSAC iterations.}

\item{conf}{\code{numeric} - confidence level.}
}
\description{
Fits a circle on XY coordinates from a set of 3D points.
}
