% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fastPointMetrics}
\alias{fastPointMetrics}
\title{Calculate metrics on point neighborhoods}
\usage{
fastPointMetrics(
  las,
  method = ptm.voxels(),
  which_metrics = ENABLED_POINT_METRICS
)
}
\arguments{
\item{las}{\code{\link[lidR:LAS]{LAS}} object.}

\item{method}{neighborhood search algorithm - currently available: \code{\link{ptm.voxels}} and \code{\link{ptm.knn}}.}

\item{which_metrics}{optional \code{character} vector - list of metrics to be calculated. Overwrites the gobally enabled metrics set with \code{fastPointMetrics.available()}.}
}
\value{
\code{LAS} object with extra fields - one for each calculated metric.
}
\description{
Get statistics for every point's neighborhood. Check out \code{\link{fastPointMetrics.available}} for information about available metrics. Neighborhood search methods are prefixed by \code{ptm}.
}
\examples{
file = system.file("extdata", "pine.laz", package="TreeLS")
tls = readTLS(file)
}
