% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gpsTimeFilter}
\alias{gpsTimeFilter}
\title{Filter points based on the \code{gpstime} field}
\usage{
gpsTimeFilter(las, from = 0, to = 1)
}
\arguments{
\item{las}{\code{\link[lidR:LAS]{LAS}} object.}

\item{from, to}{\code{numeric} - gpstime percentile thresholds (from 0 to 1) to keep points in between.}
}
\value{
\code{\link[lidR:LAS]{LAS}} object.
}
\description{
This is a simple wrapper to \code{\link[lidR:filter_poi]{filter_poi}} that takes as input relative values instead of absolute time stamps for filtering \code{LAS} object based on the gpstime. This function is particularly useful to check narrow intervals of point cloud frames from mobile scanning data.
}
