% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methods.R
\name{map.pick}
\alias{map.pick}
\title{Tree mapping algorithm: pick trees manually}
\usage{
map.pick(map = NULL, min_h = NULL, max_h = NULL)
}
\arguments{
\item{map}{optional \code{LAS} object with a tree map that beeds to be updated.}

\item{min_h, max_h}{\code{numeric} - height thresholds to filter the point cloud before plotting.}
}
\description{
This function is meant to be used inside \code{\link{treeMap}}. It opens an interactive \code{rgl} plot where the user can specify tree locations by clicking.
}
