% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{readTLS}
\alias{readTLS}
\title{Wrapper to read point clouds straight to LAS objects suitable for TLS applications}
\usage{
readTLS(file, colNames = NULL, ...)
}
\arguments{
\item{file}{object to be converted or reset}

\item{colNames}{parameter passed to \code{\link{toLAS}} whenever reading table-like files - default = \code{NULL}}

\item{...}{further arguments passed to either \code{readLAS} or \code{read.table}}
}
\value{
\code{LAS} object
}
\description{
Reads \emph{las} or \emph{laz} files with \code{\link{lidR::readLAS}}, or tries to read other file formats with \code{\link{read.table}}
}
