% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setTLS}
\alias{setTLS}
\title{Resets or creates a \code{LAS} object depending on the input's type}
\usage{
setTLS(cloud, colNames = NULL)
}
\arguments{
\item{cloud}{object to be converted or reset}

\item{colNames}{optional - only used for non-LAS objects. It states the \code{file}'s column names - if not set, only the 3 first columns will be converted to \code{LAS}}
}
\value{
\code{LAS} object
}
\description{
Resets the input whenever it is a \code{LAS} object, or generates a new \code{LAS} from a table-like input
}
