% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{shapeFit.forks}
\alias{shapeFit.forks}
\title{EXPERIMENTAL: Point cloud multiple circle fit}
\usage{
shapeFit.forks(
  dlas,
  pixel_size = 0.02,
  max_d = 0.4,
  votes_percentile = 0.7,
  min_density = 0.25,
  plot = FALSE
)
}
\arguments{
\item{dlas}{\code{\link[lidR:LAS]{LAS}} object.}

\item{pixel_size}{\code{numeric} - pixel side length to discretize the point cloud layers while performing the Hough Transform circle search.}

\item{max_d}{\code{numeric} - largest tree diameter expected in the point cloud.}

\item{votes_percentile}{\code{numeric} - use only estimates with more votes than \code{votes_percentile}.}

\item{min_density}{\code{numeric} - between 0 and 1 - minimum point density within a pixel evaluated on the Hough Transform - i.e. only \emph{dense} point clousters will undergo circle search.}

\item{plot}{\code{logical} - plot the results?}
}
\description{
Search and fit multiple 2D circles on a point cloud layer from a single tree, i.e. a forked stem segment.
}
