% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_methods.R
\name{smp.randomize}
\alias{smp.randomize}
\title{Point sampling algorithm: random sample}
\usage{
smp.randomize(p = 0.5)
}
\arguments{
\item{p}{\code{numeric} - sampling probability (from 0 to 1).}
}
\description{
This function is meant to be used inside \code{\link{tlsSample}}. It selects points randomly, returning a fraction of the input point cloud.
}
