% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stemPoints}
\alias{stemPoints}
\title{Single tree stem point classification}
\usage{
stemPoints(las, hstep = 0.5, max_radius = 0.25, hbase = c(1, 2.5),
  pixel_size = 0.025, min_density = 0.1, min_votes = 3)
}
\arguments{
\item{las}{\code{LAS} object}

\item{hstep}{...}

\item{max_radius}{...}

\item{hbase}{...}

\item{pixel_size}{...}

\item{min_density}{...}

\item{min_votes}{...}
}
\value{
\code{LAS} object
}
\description{
Classify stem points through the Hough Transform algorithm - it searches for only \emph{one} stem
}
