% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stemPoints}
\alias{stemPoints}
\title{Single tree stem point classification}
\usage{
stemPoints(las, hstep = 0.5, max_radius = 0.25, hbase = c(1, 2.5),
  pixel_size = 0.025, min_density = 0.1, min_votes = 3)
}
\arguments{
\item{las}{\code{LAS} object}

\item{hstep}{\code{numeric} - ...}

\item{max_radius}{\code{numeric} - ...}

\item{hbase}{\code{numeric} vector of length 2 - ...}

\item{pixel_size}{\code{numeric} - ...}

\item{min_density}{\code{numeric} between 0 and 1 - ...}

\item{min_votes}{\code{integer} - ...}
}
\value{
\code{LAS} object
}
\description{
Classify stem points through the Hough Transform algorithm - it searches for only \emph{one} stem
}
