% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tlsCrop}
\alias{tlsCrop}
\title{Crops a point cloud using a circle or square}
\usage{
tlsCrop(las, x, y, len, circle = T, negative = F)
}
\arguments{
\item{las}{\code{LAS} object}

\item{x, y}{X and Y center coordinates of the area to be cropped}

\item{len}{if \code{circle = TRUE}, \code{len} is the circle's radius, otherwise it is the side length of a square}

\item{circle}{if \code{TRUE}, crops a circle, otherwise a square}

\item{negative}{if \code{TRUE}, returns all points outside the specified circle/square, otherwise returns all points inside the circle/square}
}
\value{
\code{LAS} object
}
\description{
Returns a cropped point cloud of all points inside or outside specified boundaries.
}
