% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tlsRotate}
\alias{tlsRotate}
\title{Rotate point cloud to fit a horizontal ground plane}
\usage{
tlsRotate(las)
}
\arguments{
\item{las}{\code{\link[lidR:LAS]{LAS}} object.}
}
\value{
\code{\link[lidR:LAS]{LAS}} object.
}
\description{
Check for ground points and rotates the point cloud aligning its ground surface to a horizontal plane (XY).
This function is especially useful for point clouds not georeferenced or generated through mobile scanning,
which might present a tilted coordinate system.
}
