% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{treeMap}
\alias{treeMap}
\title{Map tree occurrences from TLS data}
\usage{
treeMap(las, hmin = 1, hmax = 3, hstep = 0.5, pixel = 0.025,
  rad_max = 0.25, min_den = 0.1, min_votes = 3)
}
\arguments{
\item{las}{\code{LAS} object}

\item{hmin}{...}

\item{hmax}{...}

\item{hstep}{...}

\item{pixel}{...}

\item{rad_max}{...}

\item{min_den}{...}

\item{min_votes}{...}
}
\value{
\code{LAS} object
}
\description{
Estimates tree probability regions from a point cloud based on a Hough Transform circle search
}
