% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{treeMap}
\alias{treeMap}
\title{Map tree occurrences from TLS data}
\usage{
treeMap(las, method = map.hough(), merge = 0.2, positions_only = FALSE)
}
\arguments{
\item{las}{\code{\link[lidR:LAS]{LAS}} object.}

\item{method}{tree mapping algorithm - currently available: \code{\link{map.hough}}, \code{\link{map.eigen.knn}}, \code{\link{map.eigen.voxel}} and \code{\link{map.pick}}.}

\item{merge}{\code{numeric} - parameter passed to \code{\link{treeMap.merge}}.}

\item{positions_only}{\code{logical} - if \code{TRUE} returns a 3D \code{LAS} object, otherwise returns a 2D tree map as a \code{data.table}.}
}
\value{
signed \code{LAS} or \code{data.table}.
}
\description{
Estimates tree occurrence regions from a \strong{normalized} point cloud. Tree mapping methods are prefixed by \code{map}.
}
\examples{
file = system.file("extdata", "model_boles.laz", package="TreeLS")
tls = readTLS(file)
plot(tls)

## build a 3D map of tree occurrences
map = treeMap(tls)
plot(map, color='Radii')

## get a 2D representation of the tree map
xymap = treeMap.positions(map)
head(xymap)
}
