% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{treePoints}
\alias{treePoints}
\title{Classify areas of influence for individual trees}
\usage{
treePoints(las, map, method = trp.voronoi())
}
\arguments{
\item{las}{\code{LAS} object.}

\item{map}{output from \code{\link{treeMap}} or \code{\link{treeMap.positions}}.}

\item{method}{segmentation algorithm - currently available: \code{\link{trees.voronoi}}.}
}
\value{
\code{LAS} object with a new TreeID field.
}
\description{
Classify \code{LAS} points according to the region of influence calculated from a tree map.
}
\examples{
file = system.file("extdata", "pine_plot.laz", package="TreeLS")
tls = readTLS(file)
}
