% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{treePositions}
\alias{treePositions}
\title{Get tree XY positions from a TLS-tree map}
\usage{
treePositions(las, plot = T)
}
\arguments{
\item{las}{\code{LAS} object - output from \code{\link{treeMap}}.}

\item{plot}{\code{logical} - plot the tree map?}
}
\value{
\code{data.table} of tree IDs and XY coordinates.
}
\description{
Extracts the tree positions from a \code{LAS} object tree map
}
\examples{
file = system.file("extdata", "model_boles.laz", package="TreeLS")
tls = readTLS(file)
plot(tls)

map = treeMap(tls)
plot(map, color='Radii')

xymap = treePositions(map)
head(xymap)
}
