% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_points_methods.R
\name{trp.voronoi}
\alias{trp.voronoi}
\title{Tree points algorithm: voronoi polygons.}
\usage{
trp.voronoi()
}
\description{
This function is meant to be used inside \code{\link{treePoints}}. Assign **all** points to a \code{TreeID} based on their closest \code{\link{treeMap}} coordinate.
}
